+++ tegral-di/tegral-di-test/src/main/kotlin/guru/zoroark/tegral/di/test/Utils.kt
@@ -12,7 +12,7 @@
  * limitations under the License.
  */
 
-package guru.zoroark.tegral.di
+package guru.zoroark.tegral.di.test
 
 import guru.zoroark.tegral.di.environment.Declaration
 import guru.zoroark.tegral.di.environment.EmptyQualifier
@@ -20,6 +20,10 @@ import guru.zoroark.tegral.di.environment.Identifier
 import guru.zoroark.tegral.di.environment.Qualifier
 import guru.zoroark.tegral.di.environment.ScopedSupplier
 
+/**
+ * Creates a [Declaration], automatically creating the identifier from [T] and [qualifier] and using the given
+ * [supplier]. Returns a pair with the identifier and the declaration.
+ */
 inline fun <reified T : Any> entryOf(qualifier: Qualifier = EmptyQualifier, noinline supplier: ScopedSupplier<T>) =
     Declaration(Identifier(T::class, qualifier), supplier).let {
         it.identifier to it